+++ src/condor_starter.V5/NON_POSIX.C
@@ -238,10 +238,12 @@ bsd_rusage( clock_t user, clock_t sys )
 	clock_t_to_timeval( user, bsd_rusage.ru_utime );
 	clock_t_to_timeval( sys, bsd_rusage.ru_stime );
 
-	dprintf( D_ALWAYS, "User time = %d.%06d seconds\n",
-				bsd_rusage.ru_utime.tv_sec, bsd_rusage.ru_utime.tv_usec );
-	dprintf( D_ALWAYS, "System time = %d.%06d seconds\n",
-				bsd_rusage.ru_stime.tv_sec, bsd_rusage.ru_stime.tv_usec );
+	dprintf( D_ALWAYS, "User time = %ld.%06ld seconds\n",
+				(long)bsd_rusage.ru_utime.tv_sec, 
+				(long)bsd_rusage.ru_utime.tv_usec );
+	dprintf( D_ALWAYS, "System time = %ld.%06ld seconds\n",
+				(long)bsd_rusage.ru_stime.tv_sec, 
+				(long)bsd_rusage.ru_stime.tv_usec );
 
 	return (void *)&bsd_rusage;
 }
